my first python project (calculator)
this document discusses my first Python project, which was a calculator. I decided to create a calculator as my first Python project because I wanted something simple to start with while still being able to practice using Python.

To begin, I researched the basics of Python and how to create a simple calculator program. 

Next, I wrote the code to perform the actual calculations. This involved taking the user inputs from the interface and using Python's built-in mathematical functions to perform the requested operation. I also added error handling to ensure that the program would not crash if the user entered invalid input